requirement.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
mysql==0.0.2
mysqlclient==1.4.1
pkg-resources==0.0.0
requests==2.21.0
urllib3==1.24.1
